requirements.txt
pip~=24.1.2
kafka~=1.3.5
watchdog~=5.0.2
setuptools~=70.3.0
python-dotenv~=1.0.1
